Patch from Dan Nelson to add -o switch
Adds a "-o <filename>" option, which when used with a directory argument on
the commandline, scans the directory and dumps the postscript tree to
<filename>. I factored the printing code out of print_cb and call it directly.
I also added code that recognizes when a directory on the commandline is
really a filesystem, and in that case it prints the correct freespace in the
tree.

I made some small modifications so that all error messages print to stderr
and some formatting changes.